Removed unused removeImage function in controllers with images + now using SaleController instead of KassaController
