+++ resources/views/admin/sales.blade.php
