+++ resources/views/sales/deposit.blade.php
@@ -29,7 +29,7 @@
                 </div>
 
                 <div class="row ml-0 mr-0">
-                    <form action="{{ route('kassa.storeMoney', $user->id) }}" method="post">
+                    <form action="{{ route('sales.storeMoney', $user->id) }}" method="post">
                         @csrf
                         <input type="hidden" name="amount" value="0" id="input"/>
 
@@ -38,7 +38,7 @@
                         </button>
                     </form>
 
-                    <a href="{{ route('kassa.index') }}" class="btn outlineBtn ml-1">
+                    <a href="{{ route('sales.manageAccount', $user->id) }}" class="btn outlineBtn ml-1">
                         {{ __('Annuleer') }}
                     </a>
                 </div>
