+++ resources/views/sales/index.blade.php
